backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
sqlalchemy==2.0.35
pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
pydantic==2.9.0
openai==1.50.0
pdf2image==1.17.0
aiofiles==24.1.0
torch>=2.0.0
torchvision>=0.15.0
spacy>=3.7.0
scikit-learn>=1.5.0
joblib>=1.4.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
